fix: download relay daemon on demand
