+++ tsconfig.json
@@ -6,6 +6,7 @@
   },
   "include": [
     "src",
+    "scripts",
     "test"
   ]
 }
